# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Server #1"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/users/all": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "users"
                ]
            },
            "delete": {
                "summary": "Destroy the models for the given IDs.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
        "/users": {
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
        "/login": {
            "post": {
                "summary": "Login with email and password.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "login"
                ]
            }
        }
    },
    "tags": [],
    "components": {
        "schemas": {
            "Training": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "bigint",
                        "nullable": false,
                        "example": 1717682543
                    },
                    "name": {
                        "type": "string",
                        "description": "string(255)",
                        "nullable": false,
                        "example": "string"
                    },
                    "length": {
                        "type": "number",
                        "format": "float",
                        "description": "float",
                        "nullable": false
                    },
                    "max_people": {
                        "type": "integer",
                        "description": "integer",
                        "nullable": false,
                        "example": 1057662116
                    },
                    "type": {
                        "type": "string",
                        "description": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "length",
                    "max_people",
                    "type"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "bigint",
                        "nullable": false,
                        "example": 1893736724
                    },
                    "name": {
                        "type": "string",
                        "description": "string(255)",
                        "nullable": false,
                        "example": "string"
                    },
                    "email": {
                        "type": "string",
                        "description": "string(255)",
                        "nullable": false,
                        "example": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "two_factor_secret": {
                        "type": "string",
                        "description": "text(65535)",
                        "nullable": true,
                        "example": "string"
                    },
                    "two_factor_recovery_codes": {
                        "type": "string",
                        "description": "text(65535)",
                        "nullable": true,
                        "example": "string"
                    },
                    "two_factor_confirmed_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "role": {
                        "type": "string",
                        "description": "string",
                        "default": "user",
                        "nullable": false,
                        "example": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ]
            },
            "UserController": {
                "type": "object",
                "properties": {}
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}